- in variomodel i changed the gauss model to have exponent  1.999999 (the digits are important!!).
I cheched the behaviour of the gauss model covariance function and due to numerical instabilities when computing the SPUR(smatm), which may become slightly negative and thus smatm do be no more pos definite.

To figure this out was strange and i played arounf with the digits of 1.99999.

it turned out that seven digits are ok, if it is more, it breaks down.

in fact, the SPUR of some matrix line is really slightly negative, meaning it is like -1e-9 or something..

THIS IS still to be validated.

The same occures for the shperical model.

here, it turned out that the combination of taking

cov(-1.5*vario(h)) gibes good results..

however, this is very strange..

- other changes are just clean ups..