- minor fixes in bvariogram and vario_fit.sh
but major fix in variogrammodel:

For Gauss Covariance I figured the same oddities as Fred/Thomas had with the
exponent of the gauss cova function:

(from variogrammodel)
    r2 = r**1.99999

this is odd.. if we put it to 2., C_m is no longer
pos definite.. but with 1.99999, which is in fact nearly 2. it is ok
happens not all the time but sometimes..

I did some tests with the negative definiteness of C_m

the digit can vary up to 7 digits behind the dot. With 8 digits it becomes
somehow unstable.. I took 5 digits to be sure.
To me this sounds like a DOUBLE_PRECISION / SINGLE_PRECISION Problem..

This should be figured out more precisely but dunno how..