- removed OMP parallel environment for potential calculation.
This is due to a "bug" in the OMP paradigm.
A subroutine call within the OMP parallel region fails to provide the correct memory of shared members, which are
provided by _USE modelue in the subroutin.
It turns out, that this causes data corruption for dynamic allocated fields and has the wide ranging consequence, that
subroutine calls should be done by referenceing the fields as arguments to the subroutine.

very annoying thing :(