- ADD: preselecting starting values for lambda (see below). FIX: l_bsmatm has not to be true all the time (saves some time..)
in order to have the option to preselect a starting value for our pareto problem (lambda search),
we introduce the lamnull_cri and lamnull_fpi variables which can be set by the user via
the file

'crt.lamnull'

this can contain two variables

<lamnull_cri> (double)
<lamnull_fpi> (double)

the last one may be left empty (defaults to zero than).

They are used in the subroutine blam0 to set the lammax value.

If a complex (CRI) or DC inversion is done, the variable
<lamnull_cri> is used.

For FPI we can use the variable <lamnull_fpi>, which is a NEW option.

Introduciton of a seperate value for the FPI is done because of two reasons:
- lam0 for the FPI changes usually dramatically if the
	model is changed, thus, the preselected values of lam0 for FPI may totally
	differ to the value for the CRI/DC.
- pratical reasons for synthetic study ;)

DEFAULT
If a value of zero is choosen for each one of the lamnull_*, the program chooses

SUM(diag{JTJ})/<number of parameters> * 2 /(alpha_x + alpha_y) * 5

the first is obvious, because we like to scale upon the "mean" entries of our
sensitivity matrix.

the scaling is done than to take the artificial anisotropy regularization (alpha_x and alpha_y) into account
and last not least multiply by five just to make sure we start right at the upper most boundary
for ourt pareto problem.