- FIX: GRIDBUG2: again, the border line-elements: If the numbering is somehow distorted, e.g.
FE element has node numbers
1 2 3 4
2 5 6 3
..
and the line-border-elements are like
2 1
3 2
this caused the mixed boundary value calculation to blow up, because we need
clockwise ordering of the boundary-lines

THIS IS A SERIOUS PROBLEM WITHIN THE GRID and was never checked

I introduced now a small check for self consistency of the grid, as part II of the
grid checks.

-FIX:
in fem, the lverb switch was not conform to crtomo, changed this (verbose is now
2^10=+1024)

TODO:
put the grid checks in a more intutive suproutine, function.