working version (triangular regularisation) for the decoupling of grid cells using an input file.
CRTomo looks for a file called "decouplings.dat" in the execution
directory with the following structure:

[nr of decouplings]
[cell-nr] [neighbor-nr] [dum-factor]
[cell-nr] [neighbor-nr] [dum-factor]
...

The cell-nr and neighbor-nr refer to the indices in the corresponding
fortran arrays, i.e. the element numbers, starting with one.
Surprisingly, this works with both normal and CutMcK-grids.

The dum-factor is the (floating point) factor multiplied to dum. Set it
to zero for a full decoupling. (its possible that, at the moment, this
defaults to zero).