fix grid decoupling:
previously, the grid decoupling number from the 'decouplings.dat' file
was directly used as the regularisation strength, and thus replaced the
computed regularisation parameter. For example, this total disabled
anisotropic regularisation.

This commit fixes the issue by MULTIPLYING the value from the
decouplings.dat file to the regularisation strength variable.