Rework import of individual errors:
- previously, magnitude individual errors were supposed to be ln(|Z|)
errors (although there is this strange implementation with the
normalization factors located in the last line, perhaps this can be used
to fix this issue). However, now we divide the imported |Z|-errors by the
measured resistance, which leaves us with the log(|Z|) errors.
- Now individual phase errors are read in for both complex and FPI
inversions. By turning off the error ellipsis it should be possible to
use individual magnitude errors for the complex inversion, and the
imported individual phase errors for the FPI (no tested).

If in doubt check the output of eps.ctr with some precomputed values
(e.g., 10%). Remember to take into account the log-transformation of the
error.